Linear Projection VizRank: Factor out the functionality useful for Radviz
