+++ Orange/widgets/visualize/utils/__init__.py
@@ -15,13 +15,15 @@ from AnyQt.QtWidgets import (
     QTableView, QGraphicsTextItem, QGraphicsRectItem, QGraphicsView, QDialog,
     QVBoxLayout, QLineEdit
 )
+
 from Orange.data import Variable
 from Orange.widgets import gui
 from Orange.widgets.gui import HorizontalGridDelegate, TableBarItem
 from Orange.widgets.utils.concurrent import ConcurrentMixin, TaskState
 from Orange.widgets.utils.messages import WidgetMessagesMixin
 from Orange.widgets.utils.progressbar import ProgressBarMixin
-from Orange.widgets.widget import Msg
+from Orange.widgets.settings import Setting, ContextSetting, SettingProvider
+from Orange.widgets.widget import Msg, OWComponent
 
 
 class Result(namespace):
@@ -116,6 +118,7 @@ class VizRankDialog(QDialog, ProgressBarMixin, WidgetMessagesMixin,
         self.scheduled_call = None
         self.saved_state = None
         self.done = False
+        self.pause_on_close = False
         self.saved_progress = 0
         self.scores = []
         self.add_to_model = Queue()
@@ -151,12 +154,16 @@ class VizRankDialog(QDialog, ProgressBarMixin, WidgetMessagesMixin,
         self.button = gui.button(
             self, self, "Start", callback=self.toggle, default=True)
 
+    def set_pause_on_close(self, state):
+        self.pause_on_close = state
+
     @property
     def _has_bars(self):
         return type(self).bar_length is not VizRankDialog.bar_length
 
     @classmethod
-    def add_vizrank(cls, widget, master, button_label, set_attr_callback):
+    def add_vizrank(cls, widget, master, button_label, set_attr_callback,
+                    *, auto_start=False):
         """
         Equip the widget with VizRank button and dialog, and monkey patch the
         widget's `closeEvent` and `hideEvent` to close/hide the vizrank, too.
@@ -178,9 +185,12 @@ class VizRankDialog(QDialog, ProgressBarMixin, WidgetMessagesMixin,
         # want to mess with meta-classes either.
 
         vizrank = cls(master)
+        if auto_start:
+            vizrank.set_pause_on_close(True)
         button = gui.button(
             widget, master, button_label, callback=vizrank.reshow,
             enabled=False)
+        button.pressed.connect(vizrank.start_computation)
         vizrank.selectionChanged.connect(lambda args: set_attr_callback(*args))
 
         master_close_event = master.closeEvent
@@ -394,6 +404,10 @@ class VizRankDialog(QDialog, ProgressBarMixin, WidgetMessagesMixin,
         """Code that is run after stopping the vizrank thread"""
         pass
 
+    def closeEvent(self, event):
+        if self.pause_on_close:
+            self.pause_computation()
+        super().closeEvent(event)
 
 def run_vizrank(compute_score: Callable, iterate_states: Callable,
                 saved_state: Optional[Iterable], scores: List,
@@ -568,6 +582,105 @@ class VizRankDialogAttrPair(VizRankDialog):
         item = QStandardItem(", ".join(a.name for a in attrs))
         item.setData(attrs, self._AttrRole)
         return [item]
+
+
+class VizRankDialogNAttrs(VizRankDialog):
+    n_attrs = Setting(3)
+
+    attrsSelected = Signal([])
+    _AttrRole = next(gui.OrangeUserRole)
+
+    def __init__(self, master):
+        # Add the spin box for a number of attributes to take into account.
+        VizRankDialog.__init__(self, master)
+        OWComponent.__init__(self, master)
+
+        box = gui.hBox(self)
+        self.n_attrs_spin = gui.spin(
+            box, self, None, 3, 8, label="Number of variables: ",
+            controlWidth=50, alignment=Qt.AlignRight,
+            callback=self.on_attrs_changed)
+        gui.rubber(box)
+
+        self.last_run_n_attrs = None
+        self.attr_color = master.attr_color
+        self.attrs = []
+
+    @staticmethod
+    def max_attrs():
+        return 8
+
+    def initialize(self):
+        super().initialize()
+        self.attr_color = self.master.attr_color
+        n_cont = self.max_attrs()
+        self.n_attrs_spin.setValue(min(self.n_attrs, n_cont))
+        self.n_attrs_spin.setMaximum(n_cont)
+
+    def check_preconditions(self):
+        return super().check_preconditions() \
+               and self.master.btn_vizrank.isEnabled()
+
+    def before_running(self):
+        """
+        If the number of attributes is different than
+        in the last run, reset the saved state (if it was paused).
+        """
+        if self.n_attrs != self.last_run_n_attrs:
+            self.saved_state = None
+            self.saved_progress = 0
+        if self.saved_state is None:
+            self.scores = []
+            self.rank_model.clear()
+        self.last_run_n_attrs = self.n_attrs
+
+    def on_attrs_changed(self):
+        if self.keep_running:
+            self.pause_computation()
+        if self.rank_model.rowCount() == 0:
+            self.button.setText("Start")
+            self.button.setDisabled(False)
+        elif (new_attrs := self.n_attrs_spin.value()) != self.n_attrs:
+            self.button.setText(f"Restart with {new_attrs} variables")
+            self.button.setDisabled(False)
+        else:
+            self.button.setText("Continue" if not self.done else "Finished")
+            self.button.setDisabled(self.done)
+
+    def start_computation(self):
+        self.n_attrs_spin.setValue(self.n_attrs)
+        self.n_attrs_spin.lineEdit().deselect()
+        self.rank_table.setFocus(Qt.FocusReason.OtherFocusReason)
+        super().start_computation()
+
+    def toggle(self):
+        if self.n_attrs != self.n_attrs_spin.value():
+            self.n_attrs = self.n_attrs_spin.value()
+            self.initialize()
+        super().toggle()
+
+    def closeEvent(self, event):
+        super().closeEvent(event)
+        self.n_attrs_spin.setValue(self.n_attrs)
+
+
+class CloseVizrankMixin:
+    def closeEvent(self, event):
+        if self.vizrank is not None:
+            self.vizrank.close()
+        super().closeEvent(event)
+
+    def hideEvent(self, event):
+        if self.vizrank is not None:
+            self.vizrank.hide()
+        super().hideEvent(event)
+
+    def deleteEvent(self):
+        if self.vizrank is not None:
+            self.vizrank.keep_running = False
+            self.vizrank.shutdown()
+        super().deleteEvent()
+
 
 
 class CanvasText(QGraphicsTextItem):
